Extract UnsavedChangesModelHelper
